Rename relayertest package to conformance
The new name is more accurate because it is testing not only the
relayer, but the interaction of other chains involved too.
